+++ backend/canisters/storage_bucket/impl/src/lib.rs
@@ -66,9 +66,11 @@ impl RuntimeState {
             user_count: self.data.users.len() as u64,
             file_count: file_metrics.file_count,
             blob_count: file_metrics.blob_count,
+            pending_files: file_metrics.pending_files,
             index_sync_queue_length: self.data.index_event_sync_queue.len() as u32,
+            expiration_queue_keys: file_metrics.expiration_queue_keys,
+            expiration_queue_values: file_metrics.expiration_queue_values,
             freezing_limit: self.data.freezing_limit.value.unwrap_or_default(),
-            users_migrated: self.data.users_migrated,
             stable_memory_sizes: memory::memory_sizes(),
         }
     }
@@ -83,8 +85,6 @@ struct Data {
     created: TimestampMillis,
     freezing_limit: Timestamped<Option<Cycles>>,
     rng_seed: [u8; 32],
-    #[serde(default)]
-    users_migrated: bool,
     test_mode: bool,
 }
 
@@ -98,7 +98,6 @@ impl Data {
             created: now,
             freezing_limit: Timestamped::default(),
             rng_seed: [0; 32],
-            users_migrated: true,
             test_mode,
         }
     }
@@ -130,9 +129,11 @@ pub struct Metrics {
     pub user_count: u64,
     pub file_count: u64,
     pub blob_count: u64,
+    pub pending_files: u64,
     pub index_sync_queue_length: u32,
+    pub expiration_queue_keys: u64,
+    pub expiration_queue_values: u64,
     pub freezing_limit: Cycles,
-    pub users_migrated: bool,
     pub stable_memory_sizes: BTreeMap<u8, u64>,
 }